requirements.txt
flask
python-telegram-bot[webhooks]
fastapi
uvicorn
